+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Add last ping body in MS Teams Workflows notifications (#1024)
 - Update "?rid=<uuid>" handling to allow UUIDs with uppercase letters (#1116, @wie-niet)
 - New integration: GitHub issues (#671)
+- Add support for passing Authorization header to Prometheus metrics endpoints
 
 ### Bug Fixes
 - Fix incorrect status value in Webhook integration's $JSON placeholder